# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "G4H-RMA Quant Engine",
    "description": "Enterprise Kalman + EGARCH + MCTS Pairs Trading with Multi-Agent System",
    "version": "5.0.0"
  },
  "paths": {
...
    "/api/v1/agents/start": {
      "post": {
        "summary": "Start Engine",
        "description": "Start the trading engine",
        "operationId": "start_engine_api_v1_agents_start_post",
        "parameters": [
          {
            "name": "auto_trade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Auto Trade"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/toggle-auto-trade": {
      "post": {
        "summary": "Toggle Auto-Trade",
        "description": "Toggle auto-trading mode",
        "operationId": "toggle_auto_trade_api_v1_agents_toggle_auto_trade_post",
        "parameters": [
          {
            "name": "enable",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Enable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/decisions": {
      "get": {
        "summary": "Get Decisions",
        "description": "Get recent trading decisions",
        "operationId": "get_decisions_api_v1_agents_decisions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/consensus/{pair}": {
      "get": {
        "summary": "Get Consensus",
        "description": "Get agent consensus for a pair",
        "operationId": "get_consensus_api_v1_agents_consensus__pair__get",
        "parameters": [
          {
            "name": "pair",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pair"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
...
}